+++ packages/ad/__tests__/ios/jest.config.js
@@ -1,5 +1,0 @@
-const jestConfigurator = require("@times-components/jest-configurator").default;
-
-module.exports = jestConfigurator("ios", __dirname, {
-  coverageIgnoreGlobs: ["**/webview-event-callback-setup.js"]
-});