+++ packages/article-byline/__tests__/android/article-byline-opinion.android.test.js
@@ -1,4 +1,0 @@
-import { ArticleBylineOpinion } from "../../src/article-byline";
-import shared from "../shared.native";
-
-shared(ArticleBylineOpinion);