+++ packages/article-main-comment/src/article-main-comment.web.js
@@ -64,6 +64,7 @@ class ArticlePage extends Component {
       paidContentClassName,
       isPreview,
       additionalRelatedArticlesFlag,
+      inlineRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
       latestFromSection,
@@ -86,6 +87,7 @@ class ArticlePage extends Component {
         paidContentClassName={paidContentClassName}
         isPreview={isPreview}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
+        inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}
         latestFromSectionFlag={latestFromSectionFlag}
         latestFromSection={latestFromSection}