+++ packages/article-skeleton/__tests__/common/jest.config.js
@@ -1,5 +1,0 @@
-const jestConfigurator = require("@times-components/jest-configurator").default;
-
-module.exports = jestConfigurator("common", __dirname, {
-  coverageIgnoreGlobs: ["media-aspect-ratio.js"]
-});