+++ packages/article-skeleton/__tests__/mocks.web.js
@@ -3,7 +3,7 @@ import { mockUserState } from "@times-components/user-state";
 
 export const UserState = mockUserState();
 
-jest.mock("../src/head.web", () => "Head");
+jest.mock("../src/head", () => "Head");
 
 // eslint-disable-next-line global-require
 jest.mock("@times-components/ad", () => ({