+++ packages/article-skeleton/__tests__/web/article-with-user-state.web.test.js
@@ -309,6 +309,7 @@ const renderArticle = () => (
       onTwitterLinkPress={() => {}}
       onVideoPress={() => {}}
       spotAccountId=""
+      inlineRelatedArticlesFlag
     />
   </Context.Provider>
 );