+++ packages/date-publication/__tests__/android/date-publication-bst-pm.android.test.js
@@ -1,4 +1,0 @@
-import shared from "../shared-london";
-import { dateBSTPM } from "../constants";
-
-shared(dateBSTPM);