+++ packages/date-publication/__tests__/android/date-publication-gmt-am.android.test.js
@@ -1,4 +1,0 @@
-import shared from "../shared-london";
-import { dateGMTAM } from "../constants";
-
-shared(dateGMTAM);