+++ packages/edition-slices/__tests__/ios/tile-aa.test.js
@@ -1,3 +1,0 @@
-import shared from "../tile-aa/shared-tile-aa.native";
-
-shared();