+++ packages/slice-layout/__tests__/android/sd1and4.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../sd1and4.native";
-
-shared();