+++ packages/slice-layout/__tests__/ios/sd2and2-with-style.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../sd2and2-with-style.native";
-
-shared();