+++ packages/slice-layout/__tests__/ios/vertical-list.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../vertical-list.native";
-
-shared();