+++ packages/ts-components/src/index.ts
@@ -49,3 +49,14 @@ export {
 } from './components/olympics/schedule/OlympicsSchedule';
 
 export { InfoCard } from './components/in-article-info-card/InfoCard';
+
+export { GalleryCarousel } from './components/carousel/GalleryCarousel';
+export {
+  InArticleRelatedArticles
+} from './components/in-article-related-articles/InArticleRelatedArticles';
+export {
+  formatRelatedArticles
+} from './components/in-article-related-articles/formatters';
+export {
+  InfoCardBulletPoints
+} from './components/in-article-info-card-bulletpoints/InfoCardBulletPoints';